pyproject.toml
[tool.poetry]
name = "vanna-demo"
version = "0.1.0"
description = ""
authors = ["Izhar Gavish <izharg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
vanna = {extras = ["chromadb", "openai", "postgres"], version = "^0.5.4"}
kaleido = "0.2.1"
chromadb = "^0.5.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
